feat: move project to es modules
There are various issues to overcome when using yarn pnp, ts-node and es
modules which made the move a bit more painful than it should have been.

Bluf:

- Its surprisingly complicated. Read the issue above, read the guide it
references, read the typescript post on it ... and then you might be
where I am and still only understand it 80% 😂
- Append .js to imports. This primarily applies to imports from inside
the monorepo but in some instances (fp-ts) it is also required for deps.
- ts-node doesn't play nicely with pnp ... so using the node-modules
linker for yarn.

Signed-off-by: Tim Holmes-Mitra <tim@glimpseprotocol.io>